fix(ClientState): Merged client into home
Merged client code into a state-based single page. The client page
has been removed. There is now a single page to handle spectators
and clients. When joining, you are presented with the option to
join the game. Once in the game you are able to revert to
spectating or you may also just exit the browser to leave the game.